requirements.txt
aniso8601==8.1.0
attrs==20.3.0
Authlib==0.15.2
backcall==0.2.0
Bootstrap-Flask==1.5.1
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cryptography==3.3.1
cycler==0.10.0
dash==1.18.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
decorator==4.4.2
dominate==2.6.0
dynaconf==3.1.2
Fiona==1.8.18
Flask==1.1.2
Flask-Compress==1.8.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatten-json==0.1.7
future==0.18.2
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jupyter-client==6.1.7
jupyter-core==4.7.0
kaleido==0.2.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
more-itertools==8.7.0
munch==2.5.0
nltk==3.5
numexpr==2.7.2
numpy==1.19.4
oauthlib==3.1.0
orca==1.5.4
pandas==1.2.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
prompt-toolkit==3.0.9
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyproj==3.0.0.post1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
ruamel.yaml.clib==0.2.2
ruamel.yaml.jinja2==0.2.4
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
searchtweets-v2==1.0.4
# Editable install with no version control (sentipt==0.0.1)
-e /home/philos/Downloads/Philos Data/Porfolio/Dev/sentipt
Shapely==1.7.1
six==1.15.0
SQLAlchemy==1.3.22
tables==3.6.1
textblob==0.15.3
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
tweepy==3.10.0
tweet-parser==1.13.2
unicodedata2==13.0.0.post2
urllib3==1.26.2
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
wordcloud==1.8.1
WTForms==2.3.3
yaml-1.3==0.1.0
